Checkpoint work in progress
There are two things in this commit. First, there are links to some
resources, most of which are of very high quality.

Second, there is the beginning of a mathematically rigorous approach to
shape filling based on winding numbers. However, it should be considered
a very rough draft, as it could almost certainly be formulated much
better.